Progress made on get_cells subproject!
Stephen Roller recommended parsing textgrounder logs to find out what
the geographic cells are. I've built the parsing component; I can now
load up a set of cells found in the logfile, and whether or not this
meets the full count that the log says should be present.

Scala parser combinators ftw.