get_cells now matches up cells and locations.
Yes, it's kind of wonky, and yes, when there are a full set of cells,
it will probably be quite slow, but it seems to work.

All that's needed for #8 now is to do full runs of textgrounder with the
cell models we're interested in, and then run the logs through this,
along with the wiki files.